Slightly reorganize the code
Let's try hierarchical folder structure for now
