--- resources.qrc
@@ -1,16 +1,0 @@
-<RCC>
-    <qresource prefix="/new/resource">
-        <file>trashcan.png</file>
-        <file>text-editor.png</file>
-        <file>network-wired.png</file>
-        <file>edit-find.png</file>
-        <file>process-stop.png</file>
-        <file>traffic_light_green.png</file>
-        <file>traffic_light_red.png</file>
-        <file>traffic_light_yellow.png</file>
-        <file>mono_lock.png</file>
-        <file>traffic_light_off.png</file>
-        <file>network-connected.png</file>
-        <file>network-disconnected.png</file>
-    </qresource>
-</RCC>
