+++ src/cryptdata.cpp
