+++ src/cryptdata.h
