+++ src/dialog/editdialog.cpp
