+++ src/dialog/editdialog.h
