+++ src/dialog/editdialog.ui
