+++ src/dialog/logdialog.cpp
