+++ src/dialog/logdialog.h
