+++ src/dialog/logdialog.ui
