+++ src/dialog/mainwindow.h
