+++ src/dialog/mainwindow.ui
@@ -29,8 +29,8 @@
    <string>Openconnect</string>
   </property>
   <property name="windowIcon">
-   <iconset resource="resources.qrc">
-    <normaloff>:/new/resource/mono_lock.png</normaloff>:/new/resource/mono_lock.png</iconset>
+   <iconset resource="../openconnect-gui.qrc">
+    <normaloff>:/new/resource/images/mono_lock.png</normaloff>:/new/resource/images/mono_lock.png</iconset>
   </property>
   <property name="autoFillBackground">
    <bool>false</bool>
@@ -148,7 +148,7 @@
                    <string>Edit</string>
                   </property>
                   <property name="icon">
-                   <iconset resource="resources.qrc">
+                   <iconset>
                     <normaloff>:/new/resource/text-editor.png</normaloff>:/new/resource/text-editor.png</iconset>
                   </property>
                   <property name="toolButtonStyle">
@@ -168,7 +168,7 @@
                    <string>Delete</string>
                   </property>
                   <property name="icon">
-                   <iconset resource="resources.qrc">
+                   <iconset>
                     <normaloff>:/new/resource/trashcan.png</normaloff>:/new/resource/trashcan.png</iconset>
                   </property>
                  </widget>
@@ -194,7 +194,7 @@
                  <string>Connect</string>
                 </property>
                 <property name="icon">
-                 <iconset resource="resources.qrc">
+                 <iconset>
                   <normaloff>:/new/resource/network-wired.png</normaloff>:/new/resource/network-wired.png</iconset>
                 </property>
                </widget>
@@ -214,7 +214,7 @@
                  <string>Disconnect</string>
                 </property>
                 <property name="icon">
-                 <iconset resource="resources.qrc">
+                 <iconset>
                   <normaloff>:/new/resource/process-stop.png</normaloff>:/new/resource/process-stop.png</iconset>
                 </property>
                </widget>
@@ -225,7 +225,7 @@
                  <string>View log</string>
                 </property>
                 <property name="icon">
-                 <iconset resource="resources.qrc">
+                 <iconset>
                   <normaloff>:/new/resource/edit-find.png</normaloff>:/new/resource/edit-find.png</iconset>
                 </property>
                </widget>
@@ -479,7 +479,7 @@
  </widget>
  <layoutdefault spacing="6" margin="11"/>
  <resources>
-  <include location="resources.qrc"/>
+  <include location="../openconnect-gui.qrc"/>
  </resources>
  <connections/>
 </ui>
