+++ src/gtdb.cpp
