+++ src/gtdb.h
