+++ src/images/traffic_light_green.png
