+++ src/images/traffic_light_red.png
