+++ src/images/traffic_light_yellow.png
