+++ src/main.cpp
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "mainwindow.h"
+#include "dialog/mainwindow.h"
 #include <QApplication>
 #include <QCoreApplication>
 #include <QMessageBox>
