+++ src/openconnect-gui.exe.manifest
