+++ src/openconnect-gui.ico
