+++ src/openconnect-gui.qrc
@@ -1,0 +1,16 @@
+<RCC>
+    <qresource prefix="/new/resource">
+        <file>images/trashcan.png</file>
+        <file>images/text-editor.png</file>
+        <file>images/network-wired.png</file>
+        <file>images/edit-find.png</file>
+        <file>images/process-stop.png</file>
+        <file>images/traffic_light_green.png</file>
+        <file>images/traffic_light_red.png</file>
+        <file>images/traffic_light_yellow.png</file>
+        <file>images/mono_lock.png</file>
+        <file>images/traffic_light_off.png</file>
+        <file>images/network-connected.png</file>
+        <file>images/network-disconnected.png</file>
+    </qresource>
+</RCC>
