+++ src/openconnect-gui.rc
@@ -1,7 +1,7 @@
 // ---------------------------------------------------------------------------
 // XXX: this is manually written
 // ---------------------------------------------------------------------------
-1 24 "openconnect-gui.exe.manifest"
+//1 24 "openconnect-gui.exe.manifest"
 
 // ---------------------------------------------------------------------------
 // XXX: this is autogenerated from qmake & merged here
@@ -12,7 +12,7 @@
 #  include <windows.h>
 # endif
 
-IDI_ICON1	ICON	DISCARDABLE	"logo.ico"
+IDI_ICON1	ICON	DISCARDABLE	"openconnect-gui.ico"
 
 VS_VERSION_INFO VERSIONINFO
     FILEVERSION 1,3,0,0
